GOT IT WORKING, WE GOT A JOIN TABLE NOW BABY
