+++ app/models/condition_trip.rb
@@ -1,4 +1,4 @@
-class ConditionsTrips < ActiveRecord::Base
+class ConditionTrip < ActiveRecord::Base
 
   belongs_to :trip
   belongs_to :condition
